moved class objects out of store and into utils, deleted socket
