--- client/socket.js
@@ -1,9 +1,0 @@
-import io from 'socket.io-client';
-
-const socket = io(window.location.origin);
-
-socket.on('connect', function () {
-  console.log('Connected!');
-});
-
-export default socket
