+++ client/utils/index.js
@@ -1,3 +1,4 @@
 export * from './helperfunctions'
 export * from './BSTHelperFunctions'
 export * from './drawNodeHelperFunctions'
+export * from './classObjects'
