+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "boilermaker",
+  "name": "nodeLego",
   "version": "1.0.0",
-  "description": "Some boilerplate code to get you started - get shakin'!",
+  "description": "A data structures visualization tool - engineered by Grace, Karina, Carol and Julia",
   "main": "index.js",
   "scripts": {
     "seed": "node ./seed",
@@ -41,8 +41,7 @@
     "redux": "^3.6.0",
     "redux-logger": "^2.8.1",
     "redux-thunk": "^2.2.0",
-    "sequelize": "^4.3.1",
-    "socket.io": "^2.0.3"
+    "sequelize": "^4.3.1"
   },
   "devDependencies": {
     "babel-core": "^6.22.1",
